Thorough re-word of Petoneer module, forked from sh00t2kill/petoneer_revogi_py
 - Fixed bugs (inc missing _url self method)
 - Created helper functions to translate TDS and WaterLevel values from API into human readable strings
 - Added methods to control the LEDs on fountains (as well as the pump)
 - Added internal methods to convert Petoneer API schedule times (eg: values in “section” array in JSON) to and from Python Time object.
- Implemented error handling for server errors (and also if device is offline and expected data is not returned - further work to standardise error handling and return values upon errors.

TO-DO:  Work out how the values of “watertime”, “filtertime” and “motortime” from the API device status should be interpreted to present in a user-friendly form.

TO-DO:  Further clean up error handling and return values from methods, should things not go smoothly when calling the Petoneer API

TO-DO:  To implement this module with home automation integration.